Add DomainSidPlugin unit test
Add a SECURITY hive file from a clean non-activated offline Windows 10
install that is known not to contain any security-sensitive info.

The domain SID value is not extracted because the system in not joined
into a domain, but the default workgroup name is reported correctly.

Fix permissions of the SYSTEM hive test data files.